Fix touchscreen crashes
NOTE: This does not get touch working. Only MOTION and DOWN events are
sent to a client. This is because wlc_touch_touch tries to find the
underlying view via view_under_touch which fails for UP, FRAME, CANCEL since
they do not send the pos, so we return before
wl_touch_send_up,frame,cancel can be called.