Add WIP section 7 manual for wlc
A section 7 manual is a wonderful resource for introducing library
concepts and showing rational or acting as a directed index for the
other manuals in the system.

Since this is a fairly high level manual most of it is unfinished which
hopefully overtime it gets fleshed out based on how the library is used
and the design settles.