wlc: Use eglGetPlatformDisplayEXT where available
eglGetDisplay does not know what sort of handle it is getting passed
in (a gbm device handle or an X11 display pointer) on systems without
glvnd some poking into internal X11 / gbm structures is used to figure
this out, which is somewhat dangerous but works.

On system using glvnd this does not work and eglGetDisplay returns NULL.

This commit makes wlc track the EGLNativeDisplayType pointer's
underlying type and uses eglGetPlatformDisplayEXT where available so
that the EGL implementation does not need to guess the type of the
handle it is being passed.

BugLink: https://bugzilla.redhat.com/show_bug.cgi?id=1413579
Signed-off-by: Hans de Goede <hdegoede@redhat.com>