Allow other data-source implementations
Make wlc_data_source polymorphic. This enables other data-sources
than ones provided by a client.
This can be used to manually set the clipboard data from the
compositor (and later on for x11-wayland clipboard sync).
To demonstrate this, the basic example now starts with the string
"wlc" in the clipboard.
Also add a system signal to be triggered when the data source is
changed.