Add selection retrieve to example; Small fixes
This fixes the wlc_get_selection_data (and types) implementation
and uses it in the example (ctrl+comma will try to retrieve the
selection data in plain text format print it).

Additionally this fixes various bugs. Some smaller leaks
found with valgrind as well as a critical bug where
wlc_xwm_release was called 2 times in the shutdown
process, causing a segmentation fault in x/selection.c.